+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.5.4 (tbd)
 
 - Fixed `pilet build` with `--type standalone` to end up at feed selection
+- Improved handling of server restart (#687)
 
 ## 1.5.3 (April 10, 2024)